Move recentlyViewed component to component folder
