+++ src/components/recentlyViewed/index.js
