+++ src/template/index.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { Container } from 'react-bootstrap';
 import NavBar from '../components/Navbar';
 import Footer from '../components/footer';
-import RecentlyViewed from '../recentlyViewed';
+import RecentlyViewed from '../components/recentlyViewed';
 
 export default function Template(props) {
   const { children } = props;
